Added cache with ETS table
Implemented a caching service backed by an ETS table under a GenServer.

The GenServer is used as a means to run the service under the
application umbrella and be able to query the cache concurrently.
I am not actually using the GenServer state.

The cache keeps a table of visited URLs so it is fast to retrieve
the associated URL for a given key.

The controller will still update the number of visits persistently to
the database, but it will do so asynchronously. So, if the URL is in
the cache, it's very fast to return a result.